requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
aiohttp==3.9.5
Pillow==10.4.0